fix: lint + optimize some docker
