+++ packages/app/app/composables/message/room/call/useCallDeviceSettings.ts
@@ -1,6 +1,7 @@
 import type { DeviceSection } from "@/models/message/room/call/DeviceSection";
 import type { Ref } from "vue";
 
+import { readDevices } from "@/services/message/room/liveKit/readDevices";
 import { useLiveKitStore } from "@/store/message/room/liveKit";
 import { getResultAsync, noop } from "@esposter/shared";
 
@@ -12,7 +13,7 @@ interface CallDeviceDefinition {
 
 export const useCallDeviceSettings = (definitions: CallDeviceDefinition[]) => {
   const liveKitStore = useLiveKitStore();
-  const { readDevices, switchDevice } = liveKitStore;
+  const { switchDevice } = liveKitStore;
   const deviceMap = ref(new Map<MediaDeviceKind, MediaDeviceInfo[]>());
   const deviceSections = computed<DeviceSection[]>(() =>
     definitions.map(({ kind, selectedId, title }) => ({
