+++ packages/app/app/composables/shared/useOnlineSubscribable.test.ts
@@ -7,19 +7,21 @@ import { mountSuspended } from "@nuxt/test-utils/runtime";
 import { flushPromises } from "@vue/test-utils";
 import { afterEach, beforeEach, describe, expect, test, vi } from "vitest";
 
+const goOffline = () => {
+  vi.spyOn(navigator, "onLine", "get").mockReturnValue(false);
+  window.dispatchEvent(new Event("offline"));
+};
+
+const goOnline = () => {
+  vi.spyOn(navigator, "onLine", "get").mockReturnValue(true);
+  window.dispatchEvent(new Event("online"));
+};
+
 describe(useOnlineSubscribable, () => {
   let wrapper: VueWrapper;
   let source: Ref<string>;
   let callback: ReturnType<typeof vi.fn<(value: string) => Promisable<(() => Promisable<void>) | undefined>>>;
   let cleanup: ReturnType<typeof vi.fn<() => void>>;
-  const goOffline = () => {
-    vi.spyOn(navigator, "onLine", "get").mockReturnValue(false);
-    window.dispatchEvent(new Event("offline"));
-  };
-  const goOnline = () => {
-    vi.spyOn(navigator, "onLine", "get").mockReturnValue(true);
-    window.dispatchEvent(new Event("online"));
-  };
   const mountSubscribable = async () => {
     wrapper = await mountSuspended(
       defineComponent({
