+++ packages/app/app/services/message/room/liveKit/isRemoteAudioSource.ts
@@ -1,0 +1,4 @@
+import { Track } from "livekit-client";
+
+export const isRemoteAudioSource = (source: Track.Source) =>
+  [Track.Source.Microphone, Track.Source.ScreenShareAudio].includes(source);
