+++ packages/app/app/services/message/room/liveKit/readDevices.ts
@@ -1,0 +1,3 @@
+import { Room } from "livekit-client";
+
+export const readDevices = (kind: MediaDeviceKind) => Room.getLocalDevices(kind);
